Refactor : 중복 레이아웃 제거, 공통 레이아웃으로 묶음
