+++ src/components/Slider/TopPopularPicture.jsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
 import axios from "axios";
 import SliderMultiMain from "./Slider/SliderMultiMain";
+import { WidthAreaHome } from "../../styles/Layout/Layout";
 
 export default function TopPopularPicture() {
   const [popularData, setPopularData] = useState([]);
@@ -15,8 +16,8 @@ export default function TopPopularPicture() {
   }
 
   return (
-    <>
+    <WidthAreaHome>
       <SliderMultiMain data={popularData} />
-    </>
+    </WidthAreaHome>
   );
 }
