+++ src/pages/Home.jsx
@@ -3,7 +3,6 @@ import Nav from "../components/Navbar/index";
 import Title from "../utils/Title/Title";
 import {
   CommonContainer,
-  CommonContents,
   CommonContentsHome,
   CommonContentsRight,
   CommonContentsLeft,
@@ -11,7 +10,7 @@ import {
   CommonTitleArea,
   CommonTitleTitle,
   CommonTitleText,
-} from "../components/common/Layout/Layout";
+} from "../styles/Layout/Layout";
 
 import FavoriteImage from "../assets/images/board_like_bk_32.svg";
 import SliderSingleMain from "../components/Slider/Slider/SliderSingleMain";
@@ -36,29 +35,30 @@ export default function Home() {
         <MainSlide>
           <SliderSingleMain data={data} categories={"home"} />
         </MainSlide>
-        <CommonContents>
-          <CommonContentsHome>
-            <CommonContentsLeft>
-              <CommonTitleArea>
-                <CommonTitleTitle>
-                  <img
-                    src={FavoriteImage}
-                    alt="favorite"
-                    style={{ marginRight: "10px" }}
-                  />
-                  <CommonTitleText>이번주 인기사진</CommonTitleText>
-                </CommonTitleTitle>
-                <TopPopularPicture />
-              </CommonTitleArea>
-              <HomeBoardSelect selected={selected} setSelected={setSelected} />
-              <HomeBoardSummary selected={selected} />
-              <HomeBoard selected={selected} />
-            </CommonContentsLeft>
-            <CommonContentsRight>
-              <HomeChart />
-            </CommonContentsRight>
-          </CommonContentsHome>
-        </CommonContents>
+        <CommonContentsHome>
+          {/* 홈화면 오른쪽 인기사진, 게시판 목록 */}
+          <CommonContentsLeft>
+            <CommonTitleArea>
+              <CommonTitleTitle>
+                <img
+                  src={FavoriteImage}
+                  alt="favorite"
+                  style={{ marginRight: "10px" }}
+                />
+                <CommonTitleText>이번주 인기사진</CommonTitleText>
+              </CommonTitleTitle>
+            </CommonTitleArea>
+            <TopPopularPicture />
+            <HomeBoardSelect selected={selected} setSelected={setSelected} />
+            <HomeBoardSummary selected={selected} />
+            <HomeBoard selected={selected} />
+          </CommonContentsLeft>
+
+          {/* 홈화면 왼쪽 삼대력 Top10 목록 */}
+          <CommonContentsRight>
+            <HomeChart />
+          </CommonContentsRight>
+        </CommonContentsHome>
       </CommonContainer>
     </>
   );
