+++ src/pages/NotFoundScene.jsx
@@ -1,7 +1,7 @@
 import { useLocation, Link } from "react-router-dom";
 import Title from "../utils/Title/Title";
 import Nav from "../components/Navbar/index";
-import { WidthAreaSpace } from "../styles/Layout/CommonLayout";
+import { WidthAreaSpace } from "../styles/Layout/Layout";
 import { FileBtn } from "../components/Board/NewWrite/BoardWriteStyle";
 import styled from "styled-components";
