--- src/styles/Layout/CommonLayout.jsx
@@ -1,19 +1,0 @@
-import styled from "styled-components";
-
-export const WidthAreaSpace = styled.section`
-  width: 1440px;
-  margin: 0 auto;
-  margin-top: 100px;
-`;
-
-export const WidthAreaNavSpace = styled.section`
-  max-width: 1200px;
-  width: 100%;
-  margin: 0 auto;
-`;
-
-export const WidthAreaPull = styled.section`
-  width: 100%;
-  margin: 0 auto;
-  margin-top: 150px;
-`;
